Updated Quad3 (this is to the last commit)
Added lots of comments, added a range of side and top values to check for the line (a range of 10 on each side), changed all if statements to use a boolean rather than an inequality check, wrote dead end code. P.S. the speed and time values for the turn left, right, dead end have not yet been tested.